Converting to number input types
Kinda hacky but at least gives us nicer number controls and seemingly casts the string values to numbers in the JSON:
```json
{
  "instance_type": "t2.medium",
  "name": "name",
  "instance_count": "2",
  "run_max_time": 700,
  "run_delay": 10,
  "container_name": "containerName",
  "container_url": "containerUrl"
}
```

The `instanceCount` is still a regular olde `<select>` field so that still seems to be a string. We can probably cast these out to Numbers on the server side or later in the client side (or both) when we go to submit the data to the server.